Layout list transactions and state screen

+++ app/src/main/java/com/fiz/wisecrypto/ui/screens/main/home/notification/models/TypeNotification.kt
@@ -2,6 +2,6 @@ package com.fiz.wisecrypto.ui.screens.main.home.notification.models
 
 sealed class TypeNotification {
     object Info : TypeNotification()
-    data class Transaction(val value: Int) : TypeNotification()
+    data class Transaction(val value: Double) : TypeNotification()
     data class Balance(val value: Int) : TypeNotification()
 }
